refactor(arch): Phase 2 — repositories → entities (user, session, oauth-account, api-key)
- 4개 repository를 entities로 이동 + barrel index.ts 생성
- 4개 테스트 파일 colocation (entities/__tests__/)
- eslint boundaries: legacy-infra/legacy-comp → entities 임시 허용
- 25개 파일 import path 갱신 (barrel import)

Phase 2 of FSD migration (PR B: auth-related entities)
