+++ src/__tests__/infrastructure/auth/deleteAccountAction.test.ts
@@ -11,13 +11,13 @@ jest.mock('@/shared/db/client', () => ({
 jest.mock('@/infrastructure/auth/sessionCookie', () => ({
     AUTH_SESSION_COOKIE_NAME: 'siglens_session',
 }));
-jest.mock('@/infrastructure/db/userRepository', () => ({
+jest.mock('@/entities/user', () => ({
     DrizzleUserRepository: jest.fn().mockImplementation(() => ({})),
 }));
-jest.mock('@/infrastructure/db/sessionRepository', () => ({
+jest.mock('@/entities/session', () => ({
     DrizzleSessionRepository: jest.fn().mockImplementation(() => ({})),
 }));
-jest.mock('@/infrastructure/db/oauthAccountRepository', () => ({
+jest.mock('@/entities/oauth-account', () => ({
     DrizzleOAuthAccountRepository: jest
         .fn()
         .mockImplementation(() => ({ findByUserId: jest.fn() })),
