+++ src/__tests__/infrastructure/auth/logoutAction.test.ts
@@ -11,7 +11,7 @@ jest.mock('@/shared/db/client', () => ({
 jest.mock('@/infrastructure/auth/sessionCookie', () => ({
     AUTH_SESSION_COOKIE_NAME: 'siglens_session',
 }));
-jest.mock('@/infrastructure/db/sessionRepository', () => ({
+jest.mock('@/entities/session', () => ({
     DrizzleSessionRepository: jest.fn().mockImplementation(() => ({})),
 }));
 jest.mock('@/infrastructure/auth/use-cases/logoutUser', () => ({
