+++ src/__tests__/infrastructure/auth/registerAction.test.ts
@@ -11,10 +11,10 @@ jest.mock('@/shared/db/client', () => ({
     })),
     resetDatabaseClientForTests: jest.fn(),
 }));
-jest.mock('@/infrastructure/db/sessionRepository', () => ({
+jest.mock('@/entities/session', () => ({
     DrizzleSessionRepository: jest.fn().mockImplementation(() => ({})),
 }));
-jest.mock('@/infrastructure/db/userRepository', () => ({
+jest.mock('@/entities/user', () => ({
     DrizzleUserRepository: jest.fn().mockImplementation(() => ({})),
 }));
 jest.mock('@/infrastructure/db/agreementRepository', () => ({
