+++ src/__tests__/infrastructure/chat/chatAction.test.ts
@@ -2,7 +2,7 @@ import { callAiProviderRouter } from '@/infrastructure/ai/router';
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { chatAction } from '@/infrastructure/chat/chatAction';
 import { getDatabaseClient } from '@/shared/db/client';
-import { DrizzleUserApiKeyRepository } from '@/infrastructure/db/userApiKeyRepository';
+import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
 import type {
     AnalysisResponse,
     ChatActionResult,
@@ -46,7 +46,7 @@ jest.mock('@/shared/db/client', () => ({
     getDatabaseClient: jest.fn(),
 }));
 
-jest.mock('@/infrastructure/db/userApiKeyRepository', () => ({
+jest.mock('@/entities/api-key', () => ({
     DrizzleUserApiKeyRepository: jest.fn(),
 }));
