+++ src/entities/api-key/__tests__/api.test.ts
@@ -9,7 +9,7 @@ import { encryptToken } from '@/shared/db/tokenEncryption';
 import {
     DrizzleUserApiKeyRepository,
     LlmApiKeyDecryptionFailedError,
-} from '@/infrastructure/db/userApiKeyRepository';
+} from '@/entities/api-key';
 import type { SiglensDatabase } from '@/shared/db/types';
 
 const VALID_KEY_HEX = 'b'.repeat(64);
