+++ src/entities/oauth-account/__tests__/api.test.ts
@@ -1,4 +1,4 @@
-import { DrizzleOAuthAccountRepository } from '@/infrastructure/db/oauthAccountRepository';
+import { DrizzleOAuthAccountRepository } from '@/entities/oauth-account';
 import { encryptToken } from '@/shared/db/tokenEncryption';
 import type { SiglensDatabase } from '@/shared/db/types';
