+++ src/entities/oauth-account/index.ts
@@ -1,0 +1,1 @@
+export { DrizzleOAuthAccountRepository } from './api';
