+++ src/infrastructure/auth/registerAction.ts
@@ -18,9 +18,9 @@ import { isSecureCookieEnv } from '@/infrastructure/auth/sessionCookieOptions';
 import { loginUser } from '@/infrastructure/auth/use-cases/loginUser';
 import { registerUser } from '@/infrastructure/auth/use-cases/registerUser';
 import { DrizzleAgreementRepository } from '@/infrastructure/db/agreementRepository';
-import { DrizzleSessionRepository } from '@/infrastructure/db/sessionRepository';
+import { DrizzleSessionRepository } from '@/entities/session';
 import { DrizzleTermsRepository } from '@/infrastructure/db/termsRepository';
-import { DrizzleUserRepository } from '@/infrastructure/db/userRepository';
+import { DrizzleUserRepository } from '@/entities/user';
 import { createEmailTokenStore } from '@/infrastructure/email/tokenStore';
 import { cookies } from 'next/headers';
 import { redirect } from 'next/navigation';
