+++ src/infrastructure/chat/chatAction.ts
@@ -3,8 +3,8 @@
 import { callAiProviderRouter } from '@/infrastructure/ai/router';
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { getDatabaseClient } from '@/shared/db/client';
-import { DrizzleUserApiKeyRepository } from '@/infrastructure/db/userApiKeyRepository';
-import { DrizzleUserRepository } from '@/infrastructure/db/userRepository';
+import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
+import { DrizzleUserRepository } from '@/entities/user';
 import { getUserTier } from '@/infrastructure/tier/use-cases/getUserTier';
 import type {
     AnalysisResponse,
