+++ src/infrastructure/llm/deleteApiKeyAction.ts
@@ -2,7 +2,7 @@
 
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { getDatabaseClient } from '@/shared/db/client';
-import { DrizzleUserApiKeyRepository } from '@/infrastructure/db/userApiKeyRepository';
+import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
 import { isLlmProvider, type ApiKeyActionState } from '@/domain/llm';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
