+++ src/infrastructure/llm/getRegisteredProvidersAction.ts
@@ -2,7 +2,7 @@
 
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { getDatabaseClient } from '@/shared/db/client';
-import { DrizzleUserApiKeyRepository } from '@/infrastructure/db/userApiKeyRepository';
+import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
 import type { RegisteredProvider } from '@/domain/llm';
 
 export async function getRegisteredProvidersAction(): Promise<
