+++ src/infrastructure/llm/saveApiKeyAction.ts
@@ -2,7 +2,7 @@
 
 import { getCurrentUser } from '@/infrastructure/auth/getCurrentUser';
 import { getDatabaseClient } from '@/shared/db/client';
-import { DrizzleUserApiKeyRepository } from '@/infrastructure/db/userApiKeyRepository';
+import { DrizzleUserApiKeyRepository } from '@/entities/api-key';
 import {
     isLlmProvider,
     normalizeLlmApiKey,
