+++ src/infrastructure/market/byokGate.ts
@@ -11,9 +11,9 @@ import { getDatabaseClient } from '@/shared/db/client';
 import {
     DrizzleUserApiKeyRepository,
     LlmApiKeyDecryptionFailedError,
-} from '@/infrastructure/db/userApiKeyRepository';
+} from '@/entities/api-key';
 import { getUserTier } from '@/infrastructure/tier/use-cases/getUserTier';
-import { DrizzleUserRepository } from '@/infrastructure/db/userRepository';
+import { DrizzleUserRepository } from '@/entities/user';
 import type {
     AnalysisGateError,
     AnalysisGateErrorCode,
